+++ src/main/java/com/ll/exam/sbb/question/QuestionController.java
@@ -32,7 +32,7 @@ public class QuestionController {
     }
 
     @RequestMapping("/question/detail/{id}")
-    public String detail(Model model, @PathVariable int id, AnswerForm answerForm){
+    public String detail(Model model, @PathVariable Long id, AnswerForm answerForm){
         Question question = questionService.getQuestion(id);
 
         model.addAttribute("question", question);
